+++ src/ui/components/dialogs/index.ts
@@ -6,8 +6,10 @@ export { CreateFolderDialog } from './CreateFolderDialog';
 export { RenameDialog } from './RenameDialog';
 export { DeleteConfirmDialog } from './DeleteConfirmDialog';
 export { UploadDialog } from './UploadDialog';
+export { MetadataDialog } from './MetadataDialog';
 
 export type { CreateFolderDialogProps } from './CreateFolderDialog';
 export type { RenameDialogProps } from './RenameDialog';
 export type { DeleteConfirmDialogProps } from './DeleteConfirmDialog';
 export type { UploadDialogProps } from './UploadDialog';
+export type { MetadataDialogProps, FileMetadata } from './MetadataDialog';